refactor: extract page header
